worked on file renaming and conversion of string response to object response
