+++ src/app.js
@@ -1,7 +1,7 @@
 import logger from 'morgan';
 import express from 'express';
 import cookieParser from 'cookie-parser';
-import indexRouter from './routes/index';
+import indexRouter from './routes/Authentication';
 
 const app = express();
