+++ src/middleware/middleware.js
@@ -1,4 +1,5 @@
 export const validateEmail = (email) => {
+  // eslint-disable-next-line no-useless-escape
   const mailformat = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
   return mailformat.test(email);
 };
@@ -8,22 +9,25 @@ export const validPhoneNo = (number) => {
 };
 
 export const validateUserInput = (req, res, next) => {
-  const { firstName, lastName, email, password, phoneNo } = req.body;
+  const {
+    firstName, lastName, email,
+    password, phoneNo
+  } = req.body;
 
   if (firstName.length < 3 || firstName === '') {
-    return res.status(400).send('First name must be more than three character');
+    return res.status(400).send({ messages: 'First name must be more than three character' });
   }
   if (lastName.length < 3 || lastName === '') {
-    return res.status(400).send('last name must be more than three character');
+    return res.status(400).send({ messages: 'last name must be more than three character' });
   }
   if (!validateEmail(email)) {
-    return res.status(400).send('Invalid email');
+    return res.status(400).send({ messages: 'Invalid email' });
   }
   if (password.length < 8 || password === '') {
-    return res.status(400).send('password must be more than three character');
+    return res.status(400).send({ messages: 'password must be more than three character' });
   }
   if (!validPhoneNo(phoneNo)) {
-    return res.status(400).send('Invalid phoneNo');
+    return res.status(400).send({ messages: 'Invalid phoneNo' });
   }
   return next();
 };
